Create Implement RGB cam on Arduino Uno
We are making a drone for habitat restoration using Arduino Uno, and we need to implement visible rgb cameras to detect an area from a certain height.

This code, also defines the chip select (CS) pin for the ArduCAM module (pin 10 in this case). Make sure to connect the CS pin of the ArduCAM module to pin 10 on the Arduino Uno. Additionally,  the adjusted initialization of the CS pin is given, and added a delay for stability after setting the CS pin high.