Create Implement Thermal Sensor using Arduino Uno
The code for using the MLX90614 thermal sensor with an Arduino Uno would be almost identical to the one provided for Arduino Nano. 
Make sure to include the Adafruit MLX90614 library as explained earlier for the Arduino Nano.

For wiring, you'll again need to connect the sensor's SDA pin to A4 on the Uno and SCL pin to A5 on the Uno for I2C communication.