Create Implementing LiDAR on Arduino Nano
This code continuously reads data from the LiDAR sensor and prints the distance measured in centimeters to the serial monitor. Adjust the pin number (LIDAR_SERIAL_PIN) according to your wiring.

Make sure to install the necessary libraries for your Arduino IDE to communicate with the LiDAR sensor. For the TFMini LiDAR, you might need the "TFMini" library, which you can install via the Arduino Library Manager.

This is a basic example to get you started. Depending on your specific requirements and the capabilities of your LiDAR sensor, you might need to modify and expand this code further.