Create Implementing LiDAR on Arduino Nano-Method-2
When error is being given: "Serial 1 was not declared in scope"

SoftwareSerial can be used in this case to create a software-based serial port for communication with the LiDAR sensor.

In this modified code, we use SoftwareSerial to create a software-based serial port for communication with the LiDAR sensor. The lidarSerial object is created using the lidarSerial.begin() function to initialize the serial communication.

Make sure to connect the LiDAR sensor to the appropriate pins on the Arduino Nano according to the LIDAR_TX_PIN and LIDAR_RX_PIN definitions.